Se creo Model base para refactorizar el codigo que se repite
