+++ src/main/java/com/example/taskmanager/category/Category.java
@@ -1,18 +1,11 @@
 package com.example.taskmanager.category;
 
-import com.example.taskmanager.status.Status;
+import com.example.taskmanager.ModelBase.ModelBase;
 import com.example.taskmanager.tasks.Task;
 import jakarta.persistence.*;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-import java.sql.Timestamp;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
@@ -27,11 +20,9 @@ import java.util.UUID;
 //        }
 //)
 @Entity
-@EntityListeners(AuditingEntityListener.class)
-@EnableJpaAuditing
 @SQLDelete(sql = "UPDATE Category SET deleted = true WHERE category_id=?")
 @Where(clause = "deleted = false")
-public class Category {
+public class Category extends ModelBase {
     @Id
     @GeneratedValue
     private Long categoryId;
@@ -39,46 +30,21 @@ public class Category {
     private String name;
     @Column(nullable = true, length = 2000)
     private String description;
-    @Column(updatable = false, nullable = false, unique = true, length = 36)
-    private UUID uuid;
-    @Column(updatable = false, columnDefinition = "timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP")
-    @CreatedDate
-    private Timestamp createdDate;
-    @Column(columnDefinition = "timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP")
-    @LastModifiedDate
-    private Timestamp modifiedDate;
-    @CreatedBy
-    private Integer createdBy;
-    @LastModifiedBy
-    private Integer modifiedBy;
-
-    @Column(columnDefinition = "BOOLEAN NOT NULL DEFAULT '0'")
-    private boolean deleted;
 
     @ManyToMany(mappedBy = "categories")
     private Set<Task> tasks = new HashSet<>();
-
-
-    @OneToOne(cascade = CascadeType.PERSIST)
-    @JoinColumn(name = "status_id", referencedColumnName = "statusId")
-    private Status status;
 
     public Category() {
     }
 
     public Category(UUID uuid) {
-        this.uuid = uuid;
+        super(uuid);
     }
 
     public Category(UUID uuid, String name, String description) {
-        this.uuid = uuid;
+        super(uuid);
         this.name = name;
         this.description = description;
-    }
-
-    @PrePersist
-    public void initializeUuid() {
-        this.setUuid(UUID.randomUUID());
     }
 
     public Long getCategoryId() {
@@ -105,69 +71,12 @@ public class Category {
         this.description = description;
     }
 
-    public UUID getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(UUID uuid) {
-        this.uuid = uuid;
-    }
-
-    public Timestamp getCreatedDate() {
-        return createdDate;
-    }
-
-    public void setCreatedDate(Timestamp createdDate) {
-        this.createdDate = createdDate;
-    }
-
-    public Timestamp getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Timestamp modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public Integer getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(Integer createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public Integer getModifiedBy() {
-        return modifiedBy;
-    }
-
-    public void setModifiedBy(Integer modifiedBy) {
-        this.modifiedBy = modifiedBy;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
-    }
-
     public Set<Task> getTasks() {
         return tasks;
     }
 
     public void setTasks(Set<Task> tasks) {
         this.tasks = tasks;
-    }
-
-    // Para hacer one to one
-    public Status getStatus() {
-        return status;
-    }
-
-    public void setStatus(Status status) {
-        this.status = status;
     }
 
 }
