+++ src/main/java/com/example/taskmanager/status/Status.java
@@ -1,18 +1,13 @@
 package com.example.taskmanager.status;
 
-import com.example.taskmanager.category.Category;
+import com.example.taskmanager.ModelBase.ModelBase;
 import com.example.taskmanager.tasks.Task;
 import jakarta.persistence.*;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-import java.sql.Timestamp;
 import java.util.List;
 import java.util.UUID;
 
@@ -21,7 +16,7 @@ import java.util.UUID;
 @EnableJpaAuditing
 @SQLDelete(sql = "UPDATE Status SET deleted = true WHERE status_id=?")
 @Where(clause = "deleted = false")
-public class Status {
+public class Status extends ModelBase {
     @Id
     @GeneratedValue
     private Long statusId;
@@ -29,39 +24,21 @@ public class Status {
     private String name;
     @Column(nullable = true, length = 2000)
     private String description;
-    @Column(updatable = false, nullable = false, unique = true, length = 36)
-    private UUID uuid;
-    @Column(updatable = false, columnDefinition = "timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP")
-    @CreatedDate
-    private Timestamp createdDate;
-    @Column(columnDefinition = "timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP")
-    @LastModifiedDate
-    private Timestamp modifiedDate;
-    @CreatedBy
-    private Integer createdBy;
-    @LastModifiedBy
-    private Integer modifiedBy;
-
-    @Column(columnDefinition = "BOOLEAN NOT NULL DEFAULT '0'")
-    private boolean deleted;
 
     @OneToMany(mappedBy = "status", cascade = CascadeType.PERSIST)
     private List<Task> tasks;
-
-    @OneToOne(mappedBy = "status")
-    private Category category;
 
     public Status() {
     }
 
     public Status(UUID uuid) {
-        this.uuid = uuid;
+        super(uuid);
     }
 
     public Status(UUID uuid, String name, String description) {
+        super(uuid);
         this.name = name;
         this.description = description;
-        this.uuid = uuid;
     }
 
     public Long getStatusId() {
@@ -88,53 +65,6 @@ public class Status {
         this.description = description;
     }
 
-    public UUID getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(UUID uuid) {
-        this.uuid = uuid;
-    }
-
-    public Timestamp getCreatedDate() {
-        return createdDate;
-    }
-
-    public void setCreatedDate(Timestamp createdDate) {
-        this.createdDate = createdDate;
-    }
-
-    public Timestamp getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Timestamp modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public Integer getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(Integer createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public Integer getModifiedBy() {
-        return modifiedBy;
-    }
-
-    public void setModifiedBy(Integer modifiedBy) {
-        this.modifiedBy = modifiedBy;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
-    }
 
     @PrePersist
     public void initializeUuid() {
@@ -150,11 +80,4 @@ public class Status {
         this.tasks = tasks;
     }
 
-    public Category getCategory() {
-        return category;
-    }
-
-    public void setCategory(Category category) {
-        this.category = category;
-    }
 }
