+++ src/main/java/com/example/taskmanager/tasks/Task.java
@@ -1,37 +1,27 @@
 package com.example.taskmanager.tasks;
 
+import com.example.taskmanager.ModelBase.ModelBase;
 import com.example.taskmanager.category.Category;
 import com.example.taskmanager.status.Status;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-import java.sql.Timestamp;
 import java.util.List;
 import java.util.UUID;
 
 @Entity
-@EntityListeners(AuditingEntityListener.class)
-@EnableJpaAuditing
 @SQLDelete(sql = "UPDATE Task SET deleted = true WHERE task_id=?")
 @Where(clause = "deleted = false")
-public class Task {
+public class Task extends ModelBase {
 
     @Id
     @GeneratedValue
@@ -40,21 +30,7 @@ public class Task {
     private String name;
     @Column(nullable = true, length = 2000)
     private String description;
-    @Column(updatable = false, nullable = false, unique = true, length = 36)
-    private UUID uuid;
-    @Column(updatable = false, columnDefinition = "timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP")
-    @CreatedDate
-    private Timestamp createdDate;
-    @Column(columnDefinition = "timestamp without time zone NOT NULL DEFAULT CURRENT_TIMESTAMP")
-    @LastModifiedDate
-    private Timestamp modifiedDate;
-    @CreatedBy
-    private Integer createdBy;
-    @LastModifiedBy
-    private Integer modifiedBy;
 
-    @Column(columnDefinition = "BOOLEAN NOT NULL DEFAULT '0'")
-    private boolean deleted;
 
     @ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.PERSIST})
     @JoinTable(name = "task_category",
@@ -67,15 +43,15 @@ public class Task {
     private Status status;
 
     public Task(UUID uuid) {
-        this.uuid = uuid;
+        super(uuid);
     }
 
     public Task(String name, String description, List<Category> categories, Status status, UUID uuid) {
+        super(uuid);
         this.name = name;
         this.description = description;
         this.categories = categories;
         this.status = status;
-        this.uuid = uuid;
     }
 
     public Task() {
@@ -122,56 +98,5 @@ public class Task {
         this.status = status;
     }
 
-    public UUID getUuid() {
-        return uuid;
-    }
 
-    public void setUuid(UUID uuid) {
-        this.uuid = uuid;
-    }
-
-    public Timestamp getCreatedDate() {
-        return createdDate;
-    }
-
-    public void setCreatedDate(Timestamp createdDate) {
-        this.createdDate = createdDate;
-    }
-
-    public Timestamp getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Timestamp modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public Integer getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(Integer createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public Integer getModifiedBy() {
-        return modifiedBy;
-    }
-
-    public void setModifiedBy(Integer modifiedBy) {
-        this.modifiedBy = modifiedBy;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
-    }
-
-    @PrePersist
-    public void initializeUuid() {
-        this.setUuid(UUID.randomUUID());
-    }
 }
